requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pypdf2==3.0.1
pymupdf==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
groq==0.4.0
google-generativeai==0.3.1
reportlab==4.0.7
python-dotenv==1.0.0
jinja2==3.1.2
pdfplumber==0.10.3
tabula-py==2.9.0
camelot-py==0.11.0
opencv-python==4.8.1.78
python-magic==0.4.27